third_party/WebKit/LayoutTests/inspector/elements/styles-2/add-import-rule.html

<!DOCTYPE html>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>
<style>span { color: red }</style>
<script>
function test() {

var nodeId;
var sheetId;

ElementsTestRunner.selectNodeAndWaitForStyles("styled-span", nodeSelected);

function nodeSelected(node) {
    nodeId = node.id;
    TestRunner.cssModel.matchedStylesPromise(nodeId).then(matchedStylesBefore);
}

function matchedStylesBefore(matchedResult) {
    sheetId = matchedResult.nodeStyles()[1].styleSheetId;
    TestRunner.addResult("\n== Matched rules before @import added ==\n");
    ElementsTestRunner.dumpSelectedElementStyles(true);
    TestRunner.CSSAgent.setStyleSheetText(sheetId, "@import 'data:text/css,span{color:green}';").then(sheetTextSet);
}

function sheetTextSet() {
    ElementsTestRunner.selectNodeAndWaitForStyles("styled-span", matchedStylesAfter);
}

function matchedStylesAfter() {
    TestRunner.addResult("\n== Matched rules after @import added ==\n");
    ElementsTestRunner.dumpSelectedElementStyles(true);
    TestRunner.completeTest();
}

}
</script>
<body onload="runTest()">
<span id="styled-span"></span>
